More string exchange, with multiple players allowed (choose when starting Server).  Can also enter specific IP address and port if desired.  Run Server, input number of players, then run appropriate number of Arena, making appropriate selections.  Once all clients enter a string, the server will send to each what each sent, comma-separated. This will be like the GameStateObject later on in the project.
Immediate outstanding issues with this code:
*Need to do something about transferring multiline strings.
*Need to figure out how to thread the server so that it can spawn from
the Arena process on selection of 0 in the opening prompt instead of
using a stand-alone server.  I had attempted this but ran into problems.
It may be possible to revive some of the code now that I made other
fixes and the stand-alone versions are working; I originally got rid of
the thread server so that I figure out the basic problems.
*Input-output is crude and can easily generate exceptions.
*Reading from client sockets will also benefit from a multithreaded
approach.