MP3 playback now supported in manner similar to the SoundBank, but in only one separate thread.  Requires the JavaFX library.  Java 7 has this library, but in Eclipse, it is necessary to go to Windows->Preferences->Java->Installed JREs, find your (selected) JDK, select Edit->Add External JARs, then locate and add [YOUR_JDK_PATH_HERE]/jre/lib/jfxrt.jar .
Be sure to accurately list the desired music files as the members of the
fileList array of MusicBank. MusicBank.init() should be called before
attempting to play any music.  The play() method will play a track,
looping indefinitely, by index corresponding to the media file's
position in the fileList; if another track is playing, it will be
stopped first before beginning the new track.  The stop() method simply
stops the current track.